+++ base/business/learning_units/perms.py
@@ -23,14 +23,21 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
+import datetime
+
+from base.business.institution import find_summary_course_submission_dates_for_entity_version
 from base.models import proposal_learning_unit
 from base.models.academic_year import current_academic_year, MAX_ACADEMIC_YEAR_FACULTY, MAX_ACADEMIC_YEAR_CENTRAL
 from base.models.entity import Entity
+from base.models.entity_version import find_last_entity_version_by_learning_unit_year_id
 from base.models.enums import learning_container_year_types
 from base.models.enums.entity_container_year_link_type import REQUIREMENT_ENTITY
 from base.models.enums.proposal_state import ProposalState
 from base.models.enums.proposal_type import ProposalType
+from base.models.learning_unit_year import LearningUnitYear
+from base.models.person import is_person_linked_to_entity_in_charge_of_learning_unit
 from base.models.person_entity import is_attached_entities
+from osis_common.utils.datetime import get_tzinfo, convert_date_to_datetime
 
 FACULTY_UPDATABLE_CONTAINER_TYPES = (learning_container_year_types.COURSE,
                                      learning_container_year_types.DISSERTATION,
@@ -38,14 +45,6 @@ FACULTY_UPDATABLE_CONTAINER_TYPES = (learning_container_year_types.COURSE,
 
 PROPOSAL_CONSOLIDATION_ELIGIBLE_STATES = (ProposalState.ACCEPTED.name,
                                           ProposalState.REFUSED.name)
-
-
-def is_person_linked_to_entity_in_charge_of_learning_unit(learning_unit_year, person):
-    entity = Entity.objects.filter(
-        entitycontaineryear__learning_container_year=learning_unit_year.learning_container_year,
-        entitycontaineryear__type=REQUIREMENT_ENTITY)
-
-    return is_attached_entities(person, entity)
 
 
 def _any_existing_proposal_in_epc(learning_unit_year, _):
@@ -281,3 +280,60 @@ def _negation(predicate):
         return not predicate(*args, **kwargs)
 
     return negation_method
+
+
+def can_user_view_educational_information(user, learning_unit_year_id):
+    return LearningUnitYear.objects.filter(pk=learning_unit_year_id, summary_locked=False,
+                                           attribution__summary_responsible=True,
+                                           attribution__tutor__person__user=user).exists()
+
+
+def can_user_edit_educational_information(user, learning_unit_year_id):
+    if not can_user_view_educational_information(user, learning_unit_year_id):
+        return False
+
+    submission_dates = find_educational_information_submission_dates_of_learning_unit_year(learning_unit_year_id)
+    if not submission_dates:
+        return False
+
+    now = datetime.datetime.now(tz=get_tzinfo())
+    return convert_date_to_datetime(submission_dates["start_date"]) <= now <= \
+        convert_date_to_datetime(submission_dates["end_date"])
+
+
+def find_educational_information_submission_dates_of_learning_unit_year(learning_unit_year_id):
+    entity_version = find_last_entity_version_by_learning_unit_year_id(learning_unit_year_id)
+    if entity_version is None:
+        return {}
+
+    return find_summary_course_submission_dates_for_entity_version(entity_version)
+
+
+def can_update_learning_unit_pedagogy(learning_unit_year, person):
+    """
+    Permission to edit learning unit pedagogy needs many conditions:
+        - The person must have the permission can_edit_learning_pedagogy
+        - The person must be link to the requirement entity
+        - The person can be a faculty or a central manager
+        - The person can be a tutor:
+            - The learning unit must have its flag summary_locked to false
+            - The person must have an attribution for the learning unit year
+            - The attribution must have its flag summary responsible to true.
+
+    TODO :: The complexity of the algorithm to obtain the permission makes the system unreliable.
+    TODO :: A simplification is necessary before a real use
+
+    :param learning_unit_year: LearningUnitYear
+    :param person: Person
+    :return: bool
+    """
+    if not person.user.has_perm('base.can_edit_learningunit_pedagogy'):
+        return False
+
+    if not person.is_linked_to_entity_in_charge_of_learning_unit_year(learning_unit_year):
+        return False
+
+    if person.is_faculty_manager() or person.is_central_manager():
+        return True
+
+    return can_user_edit_educational_information(person.user, learning_unit_year.id)
